pyproject.toml
[tool.poetry]
name = "wallets-tracking"
version = "0.1.0"
description = ""
authors = ["John Smith <john@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
flask = "^3.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
